Break Renderer.cpp into two files separating resource management vs rest of the code
